--- a/iree/compiler/Codegen/LLVMCPU/test/materialize_double_tiling_expert_configuration.mlir
+++ b/iree/compiler/Codegen/LLVMCPU/test/materialize_double_tiling_expert_configuration.mlir
@@ -56,5 +56,5 @@ hal.executable private @matmul_x86  {
 }
 
 //  CHECK-DAG: #[[TRANSLATION:.+]] = #iree_codegen.translation.info<"CPUDoubleTilingExpert", workload_per_wg = [64, 64]>
-//  CHECK-DAG: #[[CONFIG:.+]] =  #iree_codegen.lowering.config<tile_sizes = [{{\[}}], [288, 128, 512], [9, 32, 16]], native_vector_size = [9, 32, 16]>
+//  CHECK-DAG: #[[CONFIG:.+]] =  #iree_codegen.lowering.config<tile_sizes = [{{\[}}], [288, 128, 512], [8, 32, 16]], native_vector_size = [8, 32, 16]>
 //  CHECK:       linalg.matmul {lowering.config = #[[CONFIG]]}