@MaheshRavishankar 2022-01-26
Looks fine overall. Does this regress the miniLM perf that is being tracked. If it does, then it will just cause unnecessary pain. Would be good to check before we land this.