@hanhanW 2022-01-26
> Looks fine overall. Does this regress the miniLM perf that is being tracked. If it does, then it will just cause unnecessary pain. Would be good to check before we land this.

It's noisy, so I ran it several times. We have slight improvements in miniLM perf.

| Before (single-threaded) | After (single-threaded) | Before (multi-threaded) | After (multi-threaded) |
| ------------------------ | ----------------------- | ----------------------- | ---------------------- |
|                  82.7 ms |                 79.1 ms |                21.6 ms |                21.1 ms |