@hanhanW 2022-01-27
> LGTM! We need to port the existing verifier over to be valid for this pipeline.

Yes, this is the next thing on my plate. :)